add LazyLoad

+++ Ioc.Repository.Repositories/Models/Repository/IDataRepository.cs
@@ -2,11 +2,11 @@ using System.Collections.Generic;
 
 namespace Ioc.Repository.Repositories.Models.Repository
 {
-    public interface IDataRepository<TEntity, TDto>
+    public interface IDataRepository<TEntity, TDto> 
     {
         IEnumerable<TEntity> GetAll();
-        TEntity Get(long id);
-        TDto GetDto(long id);
+        TEntity Get(int id);
+        TDto GetDto(int id);
         void Add(TEntity entity);
         void Update(TEntity entityToUpdate, TEntity entity);
         void Delete(TEntity entity);
